Make code future proof by adding MonadFail/Semigroup instances
This patch makes the code future proof under the MonadFail and Semigroup
proposals by adding instances in a CPP-avoiding way.

Moreover, Monad instance definitions are refactored in a canonical form.

With this patch, attoparsec is warning free with GHC 8.0 under

  -Wall -Wcompat -Wnoncanonical-monad-instances -Wnoncanonical-monadfail-instances

(Compilation of attorparsec with GHCs back till 7.0.4 is warning-free as well)